+++ speechmatics/cli.py
@@ -522,15 +522,19 @@ def parse_args(args=None):
     transcribe_subparser.add_argument(
         "--enable-partials",
         default=False,
-        action="store_true"
+        action="store_true",
+        help=(
+            "Whether to return partial transcripts which can be updated "
+            "by later, final transcripts."
+        ),
     )
     transcribe_subparser.add_argument(
         "--enable-entities",
         default=False,
         action="store_true",
         help=(
-            "Whether to enable the output of ITN entities in the transcript. "
-            "Currently unsupported in Speechmatics products."
+            "Whether to output additional information about "
+            "recognised entity classes (JSON output only)."
         ),
     )
     transcribe_subparser.add_argument(
@@ -563,8 +567,21 @@ def parse_args(args=None):
         action="store_true",
         help="Shows a <sc> token where a speaker change was detected.",
     )
-    transcribe_subparser.add_argument("--max-delay", type=float)
-    transcribe_subparser.add_argument("--max-delay-mode", type=str)
+    transcribe_subparser.add_argument(
+        "--max-delay",
+        type=float,
+        help="Maximum acceptable delay before sending a piece of transcript.",
+    )
+    transcribe_subparser.add_argument(
+        "--max-delay-mode",
+        default="flexible",
+        choices=["fixed", "flexible"],
+        type=str,
+        help=(
+            "How to interpret the max-delay size if speech is in the middle "
+            "of an unbreakable entity like a number."
+        ),
+    )
     transcribe_subparser.add_argument(
         "--raw",
         metavar="ENCODING",
@@ -575,9 +592,21 @@ def parse_args(args=None):
         ),
     )
     transcribe_subparser.add_argument(
-        "--sample-rate", type=int, default=44_100)
+        "--sample-rate",
+        type=int,
+        default=44_100,
+        help="The sample rate in Hz of the input audio, if in raw format.",
+    )
     transcribe_subparser.add_argument(
-        "--chunk-size", type=int, default=1024*4)
+        "--chunk-size",
+        type=int,
+        default=1024*4,
+        help=(
+            "How much audio data, in bytes, to send to the server in each "
+            "websocket message. Larger values can increase latency, but "
+            "values which are too small create unnecessary overhead."
+        ),
+    )
     transcribe_subparser.add_argument(
         "--n-best-limit",
         type=int,